backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
scikit-learn==1.6.1
pandas==2.2.3
pdfplumber
groq
python-dotenv
